various test fixes
